Refactored scene navigation into Twitter component
